requirements.txt
Django
google-api-python-client
celery
redis
python-dotenv
django-celery-beat
djangorestframework